refactor: 공통 로직 분리
